requirements.txt
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.0.0

faiss-cpu>=1.7.4

scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0
joblib>=1.3.0

python-docx>=0.8.11
pdfplumber>=0.10.0

pyinstaller>=6.0.0